requirements.txt

argon2-cffi==20.1.0  
asgiref==3.4.1 
cffi==1.14.6
django==3.2.3
gunicorn==20.1.0
iniconfig==1.1.1
more-itertools==8.8.0
packaging==21.0
pip==21.1.2
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.4     
pytest-dotenv==0.5.2
python-dotenv==0.18.0
python_http_client==3.1.0
pytz==2021.1
pyyaml==5.4.1
sendgrid==6.0.5
setuptools==49.6.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing_extensions==3.7.4.3
werkzeug==2.0.1
wheel==0.36.2
mysqlclient==2.0.3 
google-cloud-core==1.5.0
google-cloud-storage==1.35.0

